Moved CurvePointGenerationException into package org.aksw.mlbenchmark.validation.measures.exceptions
