+++ src/main/java/org/aksw/mlbenchmark/validation/measures/exceptions/CurvePointGenerationException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.aksw.mlbenchmark.validation.measures;
+package org.aksw.mlbenchmark.validation.measures.exceptions;
 
 /**
  *
